requirements.txt
manim
google-generativeai
ipython